Remove same type cast to pass Clippy lint
Newer Clippy version lint against this, so we need to make it happy to
keep the CI light green.

+++ src/arch/metal.rs
@@ -537,9 +537,9 @@ impl Architecture for MetalArch {
                 nb_pmp,
             },
             extensions: ExtensionsCapability {
-                has_h_extension: (misa as usize & misa::H) != 0,
-                has_s_extension: (misa as usize & misa::S) != 0,
-                has_c_extension: (misa as usize & misa::C) != 0,
+                has_h_extension: (misa & misa::H) != 0,
+                has_s_extension: (misa & misa::S) != 0,
+                has_c_extension: (misa & misa::C) != 0,
                 has_sstc_extension,
                 has_zicbom_extension,
                 has_zicboz_extension,
